requirements.txt
audioread==3.0.0
discord.py==2.2.2
PyNaCl==1.5.0
python-dotenv==1.0.0
